+++ client/authority-discovery/build.rs
@@ -1,3 +1,3 @@
 fn main() {
-	prost_build::compile_protos(&["src/schema/dht.proto"], &["src/schema"]).unwrap();
+	prost_build::compile_protos(&["src/worker/schema/dht.proto"], &["src/worker/schema"]).unwrap();
 }
