+++ client/authority-discovery/src/worker/schema/dht.proto
